Moved base models under generic folder in workouts
