+++ workouts/models/generic/safe_entity.py
@@ -1,6 +1,9 @@
 from django.db import models
+from django.db.models.fields import Field
 from django.utils import timezone
 from returns.maybe import Maybe
+
+from workouts.models.generic.entity import Entity
 
 
 class SafeEntityQuerySet(models.QuerySet):
@@ -26,3 +29,20 @@ class SafeEntityManager(models.Manager):
 
     def hard_delete(self):
         return self.get_queryset().hard_delete()
+
+
+class SafeEntity(Entity):
+    deleted_at: Field = models.DateTimeField(blank=True, null=True)
+
+    objects: SafeEntityManager = SafeEntityManager(alive_only=True)
+    all_objects: SafeEntityManager = SafeEntityManager(alive_only=False)
+
+    class Meta:
+        abstract = True
+
+    def delete(self):
+        self.deleted_at = timezone.now()
+        self.save()
+
+    def hard_delete(self):
+        super().delete()
